refactor: изменить структуру проекта (перенос классов в пакеты)
