+++ src/ru/yandex/practicum/exceptions/ManagerSaveException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.manager;
+package ru.yandex.practicum.exceptions;
 
 public class ManagerSaveException extends RuntimeException {
