+++ src/ru/yandex/practicum/manager/impl/InMemoryTaskManager.java
@@ -1,5 +1,8 @@
-package ru.yandex.practicum.manager;
+package ru.yandex.practicum.manager.impl;
 
+import ru.yandex.practicum.exceptions.TimeConflictException;
+import ru.yandex.practicum.manager.HistoryManager;
+import ru.yandex.practicum.manager.TaskManager;
 import ru.yandex.practicum.tasks.*;
 import ru.yandex.practicum.util.Managers;
